move uuid file
